% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nicetable}
\alias{nicetable}
\title{Lisa's Summary Table Function}
\usage{
nicetable(df, covs, type, by = NA, warnmissby = FALSE, orderfreq = FALSE,
  labels = NA, stats = "mean_sd_median_range", statlabs = TRUE,
  tests = NA, percent = 2, perc.dec = 1, cont.dec = 2, pval.dec = 3,
  allcol = TRUE, testcol = TRUE, pvalcol = TRUE, dispmiss = FALSE,
  dispN = FALSE, printRMD = TRUE, paired = FALSE, blanks = TRUE,
  htmlTable = FALSE, color = "#EEEEEE")
}
\arguments{
\item{df}{Dataframe object name (REQUIRED).}

\item{covs}{Vector of covariates to include in table (REQUIRED).}

\item{type}{Vector indicating type of each covariates (REQUIRED).}

\item{by}{Variable to stratify by. Defaults to NA (no stratifying variable). No tests will be done.}

\item{warnmissby}{Whether to warn user that there are missing by variable values. Missing values will be excluded.
Default = FALSE.
1 = continuous, 2 = categorical,
3 = ordinal (factor levels must be set in ascending order prior to running).
If only one is listed, the type will be repeated for each covariate.}

\item{orderfreq}{If TRUE, categorical variables will be ordered by descending frequency. Default = FALSE.}

\item{labels}{Labels for covariates. Default = NA in which case variable names will be used.}

\item{stats}{Statistics to display for continuous variables (mean_sd_median_range (default), mean_sd, mean_sem, median_range, or median_iqr).
Can be a vector or length one to apply to all continuous variables.}

\item{statlabs}{If TRUE, continuous variable labels will include description of summary statistics. Default = FALSE.}

\item{tests}{Vector of tests to calculate p-values. If only one is entered it will apply to all covs.
If NA (default), no tests will be done.
Parametric ("p": t-test, chi-squared, anova),
Non-Parametric ("np": ranksum, Fisher's exact, kruskal-wallis),
T-test ("ttest"), Chi-Squared ("chisq"),
Kruskal-Wallis ("kw"), Mann-Whitney Rank Sum ("ranksum"), Anova ("anova"),
Fisher's Exact test ("fe") are curretly supported.}

\item{percent}{Should row (1) or column (2, default) percents be used?}

\item{perc.dec}{Number of decimals for percentages (categorcal variables). Default = 1.}

\item{cont.dec}{Number of decimals for continuous variable summary stats (mean, median, sd, iqr). Default = 2.}

\item{pval.dec}{Number of significant figures for p-values. Default = 3.}

\item{allcol}{Whether to diplay the "All Data" column. Default = TRUE.}

\item{testcol}{Whether to display the test column (names of tests). Default = TRUE.}

\item{pvalcol}{Should a column be included for p-values? TRUE (default) or FALSE.}

\item{dispmiss}{Whether to display number missing for continuous variables. Default = FALSE.}

\item{dispN}{Whether to display number non-missing for continuous variables. Default = FALSE.
dispN will overwrite dispmiss if both are TRUE.}

\item{printRMD}{Whether to print resulting table to Rmd via xtable. Default = TRUE.}

\item{paired}{Whether test should be paired (TRUE) or unpaired (FALSE = default). Only available for ttest and ranksum.}

\item{blanks}{Should blank rows be used as variable separators? Default = TRUE.}

\item{htmlTable}{Whether to use htmlTable package to display table (instead of xtable). Default = FALSE.}

\item{color}{Hex color to use for htmlTable output. Default = "#EEEEEE" (grey).}
}
\description{
This function creates a nice looking summary table similar
to the Mayo clinic SAS table macro. The function returns a dataframe.
The function also prints an html table by default for use in R markdown documents.
}
\keyword{lisa}
\keyword{summary}
\keyword{table}

